.claude/worktrees/peaceful-almeida/backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis[asyncio]==5.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12

# Settings
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# ML
scikit-learn==1.6.0
numpy==2.2.0
pandas==2.2.3
joblib==1.4.2

# HTTP Client (for Slack notifications)
httpx==0.28.1

# Email
# (uses stdlib smtplib)

# Utilities
python-dotenv==1.0.1

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
